requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
PyJWT==2.7.0
Werkzeug==2.3.6